# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery API",
        "description": "Food Delivery API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/restaurants/{restaurant}/menus": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get all menu items for a restaurant",
                "operationId": "433bc72d2a34df7c07090ff8aab1722e",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Create a menu item for a restaurant",
                "operationId": "7ec9bb58b33dc7fd05464a6b216f00a6",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMenuItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu item created successfully"
                    }
                }
            }
        },
        "/restaurants/{restaurant}/menus/all": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get all menu items with cursor pagination (for bulk fetch)",
                "operationId": "19f340f913d98235e078cc6683c26fcc",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 500
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/menus/{menu}": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get menu item details",
                "operationId": "ff6e3d980a47e145b19e88bae2ce5efc",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Update a menu item",
                "operationId": "883203ea44a6879d78f5f5d1a3dbf656",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMenuItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Delete a menu item",
                "operationId": "75e84dd42a0e04853f4e9cba32061128",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/menus/{menu}/availability": {
            "patch": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Update menu item availability",
                "operationId": "8f983798c262015f82df9597fc97f681",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMenuAvailabilityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu item availability updated successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreMenuItemRequest": {
                "required": [
                    "name",
                    "price"
                ],
                "properties": {
                    "menu_category_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Es Teh Manis"
                    },
                    "description": {
                        "type": "string",
                        "example": "Es teh manis segar",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "prep_time_minutes": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://example.com/image.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateMenuItemRequest": {
                "properties": {
                    "menu_category_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Es Teh Manis Updated"
                    },
                    "description": {
                        "type": "string",
                        "example": "Es teh manis segar",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "prep_time_minutes": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://example.com/image.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateMenuAvailabilityRequest": {
                "required": [
                    "is_available"
                ],
                "properties": {
                    "is_available": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}